feat(authentication): implement user provisioning and role retrieval in AuthenticationService
